# docs/swagger.yaml
basePath: /
definitions:
  models.Product:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      title:
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/models.User'
      userID:
        type: integer
    type: object
  models.User:
    properties:
      admin:
        type: boolean
      created_at:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: integer
      password:
        type: string
      products:
        items:
          $ref: '#/definitions/models.Product'
        type: array
      updated_at:
        type: string
    type: object
host: 127.0.0.1:8000
info:
  contact:
    email: soberkoder@swagger.io
    name: API Support
  description: This is a simple service for managing products by admin and users
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: JWT Product CRUD API with Authentication and Authorization
  version: "1.0"
paths:
  /products:
    get:
      consumes:
      - application/json
      description: Get all products data by admin
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Product'
      summary: Get all products
      tags:
      - products
    post:
      consumes:
      - application/json
      description: Post details of a product corresponding to the input id
      parameters:
      - description: create a product
        in: body
        name: models.Product
        required: true
        schema:
          $ref: '#/definitions/models.Product'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Product'
      summary: Post product details for the given id
      tags:
      - product
  /products/{Id}:
    delete:
      consumes:
      - application/json
      description: Delete details of a product corresponding to the input id
      parameters:
      - description: ID of the product
        in: path
        name: Id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Product'
      summary: Delete product details for a given id
      tags:
      - product
    get:
      consumes:
      - application/json
      description: Get details of a product corresponding to the input id
      parameters:
      - description: ID of the product
        in: path
        name: Id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Product'
      summary: Get product details for the given id
      tags:
      - product
    put:
      consumes:
      - application/json
      description: Update details of a product corresponding to the input id
      parameters:
      - description: ID of the product
        in: path
        name: Id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Product'
      summary: Update product for the given id
      tags:
      - product
swagger: "2.0"